% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMatrix.R
\name{CompareMatrix}
\alias{CompareMatrix}
\title{Comparaison de matrice 9x9}
\usage{
CompareMatrix(grille1, grille2)
}
\arguments{
\item{grille1}{une grille 9x9}

\item{grille2}{une grille 9x9}
}
\value{
booléen TRUE si grille1 est similaire à grille2
}
\description{
Comparaison de matrice 9x9
}
\examples{
m<-matrix(1:9,nrow=9,ncol=9)
CompareMatrix(m,m)
}
\author{
Camille Carrette Jean Charles Delmas
}
