% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genGrilleIncomplete.R
\name{genGrilleIncomplete}
\alias{genGrilleIncomplete}
\title{Génère une matrice incomplete}
\usage{
genGrilleIncomplete(grille, diff)
}
\arguments{
\item{grille}{une grille complete pour sudoku}

\item{diff}{entier indiquant la difficulté du sudoku}
}
\value{
la grille incomplete
}
\description{
Génère une matrice incomplete
}
\examples{
m<-matrix(1:9,rows=9,cols=9)
genGrilleIncomplete(m,1)

m<-genGrilleComplete()
genGrilleIncomplete(m,2)
}
\author{
Camille Carrette Jean Charles Delmas
}
